requirements.txt
discord.py==2.3.1
requests==2.32.2
ffmpeg-python==0.2.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.23
# selenium==4.15.2
yt-dlp==2024.7.16
pydantic